Move PrefabState/IPrefabStateInfo to common/prefab and fix OptimizationPolicy
- Move PrefabState enum and IPrefabStateInfo interface from
  common/node.ts to common/prefab/prefab-info.ts where they belong
- Fix OptimizationPolicy enum: AUTO=0, SINGLE_INSTANCE=1, MULTI_INSTANCE=2
  (previously AUTO and SINGLE_INSTANCE were both 0)
- Update imports in prefab/utils.ts to use common re-export
